+++ pkg/collector/corechecks/snmp/internal/profile/profile_yaml.go
@@ -60,31 +60,11 @@ func loadYamlProfiles() (ProfileConfigMap, error) {
 	}
 	log.Warnf("userProfiles: %v", userProfiles)
 
-	//pConfig, err := getDefaultProfilesDefinitionFiles()
-	//if err != nil {
-	//	return nil, fmt.Errorf("failed to get default profile definitions: %s", err)
-	//}
-	defaultExpandedProfiles, err := loadProfilesV2(defaultProfiles, nil)
+	profiles, err := resolveProfiles(defaultProfiles, userProfiles)
 	if err != nil {
-		return nil, fmt.Errorf("failed to load default profiles: %s", err)
+		return nil, err
 	}
-	userExpandedProfiles, err := loadProfilesV2(userProfiles, defaultProfiles)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load default profiles: %s", err)
-	}
-	profiles := ProfileConfigMap{}
-	for key, val := range defaultExpandedProfiles {
-		if strings.HasPrefix(key, "_") {
-			continue
-		}
-		profiles[key] = val
-	}
-	for key, val := range userExpandedProfiles {
-		if strings.HasPrefix(key, "_") {
-			continue
-		}
-		profiles[key] = val
-	}
+
 	SetGlobalProfileConfigMap(profiles)
 	return profiles, nil
 }
@@ -164,38 +144,6 @@ func getProfilesDefinitionFilesV2(profilesFolder string) (ProfileConfigMap, erro
 	return profiles, nil
 }
 
-func loadProfilesV2(pConfig ProfileConfigMap, defaultProfiles ProfileConfigMap) (ProfileConfigMap, error) {
-	profiles := make(ProfileConfigMap, len(pConfig))
-
-	for name, profConfig := range pConfig {
-		//if profConfig.DefinitionFile != "" {
-		//	profDefinition, err := readProfileDefinition(profConfig.DefinitionFile)
-		//	if err != nil {
-		//		log.Warnf("failed to read profile definition `%s`: %s", name, err)
-		//		continue
-		//	}
-		//
-		//}
-
-		err := recursivelyExpandBaseProfilesV2(name, &profConfig.Definition, profConfig.Definition.Extends, []string{}, pConfig, defaultProfiles)
-		if err != nil {
-			log.Warnf("failed to expand profile `%s`: %s", name, err)
-			continue
-		}
-		//profConfig.Definition = *profDefinition
-		profiledefinition.NormalizeMetrics(profConfig.Definition.Metrics)
-		errors := configvalidation.ValidateEnrichMetadata(profConfig.Definition.Metadata)
-		errors = append(errors, configvalidation.ValidateEnrichMetrics(profConfig.Definition.Metrics)...)
-		errors = append(errors, configvalidation.ValidateEnrichMetricTags(profConfig.Definition.MetricTags)...)
-		if len(errors) > 0 {
-			log.Warnf("validation errors: %s", strings.Join(errors, "\n"))
-			continue
-		}
-		profiles[name] = profConfig
-	}
-	return profiles, nil
-}
-
 func loadProfiles(pConfig ProfileConfigMap) (ProfileConfigMap, error) {
 	profiles := make(ProfileConfigMap, len(pConfig))
 
@@ -256,52 +204,6 @@ func resolveProfileDefinitionPath(definitionFile string) string {
 func getProfileConfdRoot(profileFolderName string) string {
 	confdPath := config.Datadog.GetString("confd_path")
 	return filepath.Join(confdPath, "snmp.d", profileFolderName)
-}
-
-func recursivelyExpandBaseProfilesV2(parentExtend string, definition *profiledefinition.ProfileDefinition, extends []string, extendsHistory []string, profiles ProfileConfigMap, defaultProfiles ProfileConfigMap) error {
-	//parentBasePath := filepath.Base(parentPath)
-	for _, extendEntry := range extends {
-		// Skip non yaml profiles
-		if !strings.HasSuffix(extendEntry, ".yaml") {
-			continue
-		}
-		extendEntry = extendEntry[:len(extendEntry)-len(".yaml")]
-
-		var baseDefinition *profiledefinition.ProfileDefinition
-		// User profile can extend default profile by extending the default profile.
-		// If the extend entry has the same name as the profile name, we assume the extend entry is referring to a default profile.
-		if extendEntry == parentExtend {
-			//extendEntry = filepath.Join(getProfileConfdRoot(defaultProfilesFolder), extendEntry)
-			profile, ok := defaultProfiles[extendEntry]
-			if !ok {
-				return fmt.Errorf("extend does not exist: `%s`", extendEntry)
-			}
-			baseDefinition = &profile.Definition
-		} else {
-			profile, ok := profiles[extendEntry]
-			if !ok {
-				profile, ok = defaultProfiles[extendEntry]
-				if !ok {
-					return fmt.Errorf("extend does not exist: `%s`", extendEntry)
-				}
-			}
-			baseDefinition = &profile.Definition
-		}
-		for _, extend := range extendsHistory {
-			if extend == extendEntry {
-				return fmt.Errorf("cyclic profile extend detected, `%s` has already been extended, extendsHistory=`%v`", extendEntry, extendsHistory)
-			}
-		}
-
-		mergeProfileDefinition(definition, baseDefinition)
-
-		newExtendsHistory := append(common.CopyStrings(extendsHistory), extendEntry)
-		err := recursivelyExpandBaseProfilesV2(extendEntry, definition, baseDefinition.Extends, newExtendsHistory, profiles, defaultProfiles)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
 }
 
 func recursivelyExpandBaseProfiles(parentPath string, definition *profiledefinition.ProfileDefinition, extends []string, extendsHistory []string) error {
